Moves AppStateProvider to separate directory
